pyproject.toml
[project]
name = "remembertheworld"
version = "0.1.0"
description = "This is a sample Python project for the remembertheworld website"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "feedparser>=6.0.12",
    "python-dotenv>=1.2.1",
]
